fix links in the docs
we take the first step to centralize all the links into two files:
_external_link_url and _internal_link_url.
The reason we have two is that internal link reference is relative to the
current folder of the docs, which causes the links defined for upgrade.md
is not usable for reference/task.md, for example.

The effort to centralize all the links are not finished, we might want to
dedicate some time to finish the process, but not in this story.

[#161729119]

Signed-off-by: Jesse Alford <jalford@pivotal.io>