feat: update image from bionic to jammy
Bionic hits EoL in Apr 2023, so i'm switching us to jammy (the
latest LTS) in advance to prep for this.  Also, because bionic's
python3 (3.6) was deprecated and giving customers warnings when
using cli tools.

Note: in pip 22.3 the "--use-feature=2020-resolver" became default
behavior, so we don't need to specify it anymore.